refactor: extract the logic to load/store software settings
