+++ dinstaller-lib/src/store.rs
@@ -1,8 +1,9 @@
+mod software;
 mod users;
 
-use crate::install_settings::{InstallSettings, SoftwareSettings, StorageSettings};
-use crate::software::SoftwareClient;
+use crate::install_settings::{InstallSettings, StorageSettings};
 use crate::storage::StorageClient;
+use crate::store::software::SoftwareStore;
 use crate::store::users::UsersStore;
 use std::{default::Default, error::Error};
 
@@ -11,7 +12,7 @@ use std::{default::Default, error::Error};
 /// This struct uses the default connection built by [connection function](super::connection).
 pub struct Store<'a> {
     users: UsersStore<'a>,
-    software_client: SoftwareClient<'a>,
+    software: SoftwareStore<'a>,
     storage_client: StorageClient<'a>,
 }
 
@@ -19,20 +20,16 @@ impl<'a> Store<'a> {
     pub fn new() -> Result<Self, zbus::Error> {
         Ok(Self {
             users: UsersStore::new(super::connection()?)?,
-            software_client: SoftwareClient::new(super::connection()?)?,
+            software: SoftwareStore::new(super::connection()?)?,
             storage_client: StorageClient::new(super::connection()?)?,
         })
     }
 
     /// Loads the installation settings from the D-Bus service
     pub fn load(&self) -> Result<InstallSettings, Box<dyn Error>> {
-        let product = self.software_client.product()?;
-
         let settings = InstallSettings {
             storage: Default::default(),
-            software: SoftwareSettings {
-                product: Some(product),
-            },
+            software: self.software.load()?,
             user: self.users.load()?,
         };
         Ok(settings)
@@ -40,16 +37,9 @@ impl<'a> Store<'a> {
 
     /// Stores the given installation settings in the D-Bus service
     pub fn store(&self, settings: &InstallSettings) -> Result<(), Box<dyn Error>> {
-        self.store_software_settings(&settings.software)?;
+        self.software.store(&settings.software)?;
         self.users.store(&settings.user)?;
         self.store_storage_settings(&settings.storage)?;
-        Ok(())
-    }
-
-    fn store_software_settings(&self, settings: &SoftwareSettings) -> Result<(), Box<dyn Error>> {
-        if let Some(product) = &settings.product {
-            self.software_client.select_product(product)?;
-        }
         Ok(())
     }
